feat (notifiers): added backendUrlNotifier and put notifiers in their own directory to clean the project
