+++ mobile/lib/core/notifiers/backend_address_notifier.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class BackendAddressNotifier extends ChangeNotifier {
+  String? _backendAddress;
+
+  String? get backendAddress => _backendAddress;
+
+  void setBackendAddress(String? newAddress) {
+    if (_backendAddress == newAddress) return;
+    _backendAddress = newAddress;
+    notifyListeners();
+  }
+}
