+++ mobile/lib/core/notifiers/locale_notifier.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class LocaleNotifier extends ChangeNotifier {
+  Locale _locale = const Locale('en');
+
+  Locale get locale => _locale;
+
+  void setLocale(Locale newLocale) {
+    if (_locale == newLocale) return;
+    _locale = newLocale;
+    notifyListeners();
+  }
+}
